# pkg/gen/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/pb/todo.proto",
    "version": "1"
  },
  "tags": [
    {
      "name": "TodoSVC"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/todos": {
      "get": {
        "summary": "Return the todos",
        "description": "Return the todos",
        "operationId": "TodoSVC_GetTodos",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/GetTodosOutput"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "todos"
        ]
      },
      "post": {
        "summary": "Create a new todo",
        "description": "Create a new todo. Returns created todo.",
        "operationId": "TodoSVC_CreateTodo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CreateTodoOutput"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateTodoInput"
            }
          }
        ],
        "tags": [
          "todos"
        ]
      },
      "put": {
        "summary": "Update the todo",
        "description": "Update the todo. Returns an updated todo.",
        "operationId": "TodoSVC_UpdateTodo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/UpdateTodoOutput"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateTodoInput"
            }
          }
        ],
        "tags": [
          "todos"
        ]
      }
    },
    "/v1/todos/*": {
      "get": {
        "summary": "Returns a todo",
        "description": "Returns a todo by id.",
        "operationId": "TodoSVC_GetTodo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/GetTodoOutput"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "todos"
        ]
      },
      "delete": {
        "summary": "Delete the todo",
        "description": "Delete the todo.",
        "operationId": "TodoSVC_DeleteTodo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "todos"
        ]
      }
    }
  },
  "definitions": {
    "CreateTodoInput": {
      "type": "object",
      "properties": {
        "todo": {
          "$ref": "#/definitions/Todo"
        }
      }
    },
    "CreateTodoOutput": {
      "type": "object",
      "properties": {
        "todo": {
          "$ref": "#/definitions/Todo"
        }
      }
    },
    "GetTodoOutput": {
      "type": "object",
      "properties": {
        "todo": {
          "$ref": "#/definitions/Todo"
        }
      }
    },
    "GetTodosOutput": {
      "type": "object",
      "properties": {
        "todos": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/Todo"
          }
        }
      }
    },
    "Todo": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "UpdateTodoInput": {
      "type": "object",
      "properties": {
        "todo": {
          "$ref": "#/definitions/Todo"
        }
      }
    },
    "UpdateTodoOutput": {
      "type": "object",
      "properties": {
        "todo": {
          "$ref": "#/definitions/Todo"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}